+++ src/replay_gpx.html
@@ -11,7 +11,6 @@
         <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css" />
         <link rel="stylesheet" href="./css/main.css" />
 
-        <script src="https://unpkg.com/leaflet/dist/leaflet.js"></script>
         <script src='https://unpkg.com/@turf/turf@6/turf.min.js'></script>
         <script src="./js/replay.js" type="module"></script>
     </head>
